refactor: 이벤트 헨들러 폴더 분리
